rustc: Fix enum alignment

+++ src/test/run-pass/tag-align-shape-u32.rs
@@ -2,7 +2,7 @@
 // See issue #1535
 
 enum a_tag {
-    a_tag(u64)
+    a_tag(u32)
 }
 
 type t_rec = {
@@ -11,7 +11,7 @@ type t_rec = {
 };
 
 fn main() {
-    let x = {c8: 22u8, t: a_tag(44u64)};
+    let x = {c8: 22u8, t: a_tag(44u32)};
     let y = #fmt["%?", x];
     #debug["y = %s", y];
     assert y == "(22, a_tag(44))";
